requirements.txt
fastapi==0.135.3
uvicorn==0.44.0
python-multipart==0.0.24
mlflow==3.10.1
dvc==3.67.1
torch==2.11.0
torchvision==0.26.0
scikit-learn==1.5.2
pillow==12.2.0
tqdm==4.67.3
numpy>=1.24.0
matplotlib>=3.7.0
psutil>=5.9.0
prometheus-client>=0.17.0
